refactor(backups/flr): extract flr methods to their file
